Added support for multi structure pack
This commit is thanks to Fond and Win Epic Fin It allows for things to be nametaged and for us to build multiple packs in one file
