--- lookups/armor_stand.entity.json
@@ -1,62 +1,0 @@
-{
-  "format_version": "1.10.0",
-  "minecraft:client_entity": {
-    "description": {
-      "identifier": "minecraft:armor_stand",
-      "min_engine_version": "1.8.0",
-      "materials": {
-        "default": "armor_stand",
-
-        "ghost_blocks": "entity_alphablend" //Makes ghost_blocks render semi_transparent
-
-      },
-      "textures": {
-        "default": "textures/entity/armor_stand",
-        "ghost_blocks": "textures/entity/ghost_blocks"
-      },
-      "animations": {
-        "default_pose": "animation.armor_stand.default_pose",
-        "no_pose": "animation.armor_stand.no_pose",
-        "solemn_pose": "animation.armor_stand.solemn_pose",
-        "athena_pose": "animation.armor_stand.athena_pose",
-        "brandish_pose": "animation.armor_stand.brandish_pose",
-        "honor_pose": "animation.armor_stand.honor_pose",
-        "entertain_pose": "animation.armor_stand.entertain_pose",
-        "salute_pose": "animation.armor_stand.salute_pose",
-        "riposte_pose": "animation.armor_stand.riposte_pose",
-        "zombie_pose": "animation.armor_stand.zombie_pose",
-        "cancan_a_pose": "animation.armor_stand.cancan_a_pose",
-        "cancan_b_pose": "animation.armor_stand.cancan_b_pose",
-        "hero_pose": "animation.armor_stand.hero_pose",
-        "wiggle": "animation.armor_stand.wiggle",
-        "controller.pose": "controller.animation.armor_stand.pose",
-        "controller.wiggling": "controller.animation.armor_stand.wiggle",
-        // Ghost blocks animations
-        "scale": "animation.armor_stand.ghost_blocks.scale" // For scaling the blocks to full size.		
-      },
-      "scripts": {
-        "initialize": [
-          "variable.armor_stand.pose_index = 0;",
-          "variable.armor_stand.hurt_time = 0;"
-        ],
-        "animate": [
-          "controller.pose",
-          "controller.wiggling",
-          // Ghost blocks animations
-          "scale" // For scaling the blocks to full size.
-        ]
-      },
-      "geometry": {
-
-        "default": "geometry.armor_stand.larger_render", //Changes vanilla geometry to render further than it should
-        "ghost_blocks": "geometry.armor_stand.ghost_blocks"
-
-      },
-      "render_controllers": [
-        "controller.render.armor_stand",
-        "controller.render.armor_stand.ghost_blocks" // For rendering the blocks
-      ],
-      "enable_attachables": true
-    }
-  }
-}
