--- lookups/armor_stand.ghost_blocks.render_controllers.json
@@ -1,16 +1,0 @@
-{
-	"format_version": "1.8.0",
-	"render_controllers": {
-		"controller.render.armor_stand.ghost_blocks": {
-			"geometry": "Geometry.ghost_blocks",
-			"materials": [
-				{
-					"*": "Material.ghost_blocks"
-				}
-			],
-			"textures": [
-				"Texture.ghost_blocks"
-			]
-		}
-	}
-}
